PSR-251 adding the general messages and adjusting the tests accordingly + moving the relative tests to common
