+++ conf/messages.en
@@ -690,16 +690,27 @@ fileUploadTooManyRows.heading = Your member details file has too many rows
 fileUploadTooManyRows.paragraph1 = We cannot accept the file you uploaded as it contains too many rows. The maximum number of rows it can have is 300.
 fileUploadTooManyRows.paragraph2 = Remove the additional rows and upload it again, or upload a different file.
 
-partnershipRecipientUtr.title = Do you know the recipient’s Unique Taxpayer Reference (UTR)?
-partnershipRecipientUtr.heading = Do you know {0}’s Unique Taxpayer Reference (UTR)?
-partnershipRecipientUtr.yes.conditional = What is {0}’s UTR?
-partnershipRecipientUtr.yes.conditional.error.required = Enter the recipient’s UTR
-partnershipRecipientUtr.yes.conditional.error.invalid = Enter a Unique Taxpayer Reference that is 10 numbers, like 1234567890 or 12345 67890
-partnershipRecipientUtr.no.conditional = Tell us why you cannot provide {0}’s UTR
-partnershipRecipientUtr.no.conditional.error.required = Enter the reason you cannot provide the recipient’s UTR
-partnershipRecipientUtr.no.conditional.error.invalid = The reason you cannot provide the recipient’s UTR must only include letters a to z, numbers 0 to 9, apostrophes, ‘atʼ signs (@), commas, full stops, hyphens, slashes, speech marks and spaces
-partnershipRecipientUtr.no.conditional.error.length = The reason you cannot provide the recipient’s UTR must be 160 characters or fewer
-partnershipRecipientUtr.error.required = Select yes if you know the recipient’s Unique Taxpayer Reference (UTR)
+loanRecipient.partnershipRecipientUtr.title = Do you know the recipient’s Unique Taxpayer Reference (UTR)?
+loanRecipient.partnershipRecipientUtr.heading = Do you know {0}’s Unique Taxpayer Reference (UTR)?
+loanRecipient.partnershipRecipientUtr.yes.conditional = What is {0}’s UTR?
+loanRecipient.partnershipRecipientUtr.yes.conditional.error.required = Enter the recipient’s UTR
+loanRecipient.partnershipRecipientUtr.yes.conditional.error.invalid = Enter a Unique Taxpayer Reference that is 10 numbers, like 1234567890 or 12345 67890
+loanRecipient.partnershipRecipientUtr.no.conditional = Tell us why you cannot provide {0}’s UTR
+loanRecipient.partnershipRecipientUtr.no.conditional.error.required = Enter the reason you cannot provide the recipient’s UTR
+loanRecipient.partnershipRecipientUtr.no.conditional.error.invalid = The reason you cannot provide the recipient’s UTR must only include letters a to z, numbers 0 to 9, apostrophes, ‘atʼ signs (@), commas, full stops, hyphens, slashes, speech marks and spaces
+loanRecipient.partnershipRecipientUtr.no.conditional.error.length = The reason you cannot provide the recipient’s UTR must be 160 characters or fewer
+loanRecipient.partnershipRecipientUtr.error.required = Select yes if you know the recipient’s Unique Taxpayer Reference (UTR)
+
+landOrPropertySeller.partnershipRecipientUtr.title = Do you know the recipient’s Unique Taxpayer Reference (UTR)?
+landOrPropertySeller.partnershipRecipientUtr.heading = Do you know {0}’s Unique Taxpayer Reference (UTR)?
+landOrPropertySeller.partnershipRecipientUtr.yes.conditional = What is {0}’s UTR?
+landOrPropertySeller.partnershipRecipientUtr.yes.conditional.error.required = Enter the recipient’s UTR
+landOrPropertySeller.partnershipRecipientUtr.yes.conditional.error.invalid = Enter a Unique Taxpayer Reference that is 10 numbers, like 1234567890 or 12345 67890
+landOrPropertySeller.partnershipRecipientUtr.no.conditional = Tell us why you cannot provide {0}’s UTR
+landOrPropertySeller.partnershipRecipientUtr.no.conditional.error.required = Enter the reason you cannot provide the recipient’s UTR
+landOrPropertySeller.partnershipRecipientUtr.no.conditional.error.invalid = The reason you cannot provide the recipient’s UTR must only include letters a to z, numbers 0 to 9, apostrophes, ‘atʼ signs (@), commas, full stops, hyphens, slashes, speech marks and spaces
+landOrPropertySeller.partnershipRecipientUtr.no.conditional.error.length = The reason you cannot provide the recipient’s UTR must be 160 characters or fewer
+landOrPropertySeller.partnershipRecipientUtr.error.required = Select yes if you know the recipient’s Unique Taxpayer Reference (UTR)
 
 recipientSponsoringEmployerConnectedParty.title = Is the recipient a sponsoring employer or a connected party?
 recipientSponsoringEmployerConnectedParty.heading = Is {0} a sponsoring employer or a connected party?
